requirements.txt
streamlit>=1.28.0
torch>=1.13.0,<2.0.0
torchvision>=0.14.0,<0.16.0
diffusers>=0.16.0,<0.21.0
transformers>=4.21.0,<4.25.0
accelerate>=0.16.0,<0.21.0
Pillow>=9.0.0
safetensors>=0.3.0
huggingface_hub>=0.10.0,<0.16.0